Fixed SwiftLint errors and split files up.
