+++ Sources/CodeEditSourceEditor/CodeEditSourceEditor/CodeEditSourceEditor.swift
@@ -263,15 +263,16 @@ public struct CodeEditSourceEditor: NSViewControllerRepresentable {
     /// Update the parameters of the controller.
     /// - Parameter controller: The controller to update.
     func updateControllerParams(controller: TextViewController) {
+        updateTextProperties(controller)
+        updateEditorProperties(controller)
+        updateThemeAndLanguage(controller)
+        updateHighlighting(controller)
+    }
+
+    private func updateTextProperties(_ controller: TextViewController) {
         if controller.font != font {
             controller.font = font
         }
-
-        controller.wrapLines = wrapLines
-        controller.useThemeBackground = useThemeBackground
-        controller.lineHeightMultiple = lineHeight
-        controller.editorOverscroll = editorOverscroll
-        controller.contentInsets = contentInsets
 
         if controller.isEditable != isEditable {
             controller.isEditable = isEditable
@@ -280,14 +281,14 @@ public struct CodeEditSourceEditor: NSViewControllerRepresentable {
         if controller.isSelectable != isSelectable {
             controller.isSelectable = isSelectable
         }
+    }
 
-        if controller.language.id != language.id {
-            controller.language = language
-        }
-
-        if controller.theme != theme {
-            controller.theme = theme
-        }
+    private func updateEditorProperties(_ controller: TextViewController) {
+        controller.wrapLines = wrapLines
+        controller.useThemeBackground = useThemeBackground
+        controller.lineHeightMultiple = lineHeight
+        controller.editorOverscroll = editorOverscroll
+        controller.contentInsets = contentInsets
 
         if controller.indentOption != indentOption {
             controller.indentOption = indentOption
@@ -304,7 +305,19 @@ public struct CodeEditSourceEditor: NSViewControllerRepresentable {
         if controller.useSystemCursor != useSystemCursor {
             controller.useSystemCursor = useSystemCursor
         }
+    }
 
+    private func updateThemeAndLanguage(_ controller: TextViewController) {
+        if controller.language.id != language.id {
+            controller.language = language
+        }
+
+        if controller.theme != theme {
+            controller.theme = theme
+        }
+    }
+
+    private func updateHighlighting(_ controller: TextViewController) {
         if !areHighlightProvidersEqual(controller: controller) {
             controller.setHighlightProviders(highlightProviders)
         }
